fixes for latest gt master

+++ src/atlas/interpolation/method/PointSet.cc
@@ -48,14 +48,14 @@ PointSet::PointSet( Mesh& mesh ) {
         std::vector<PointIndex3::Value> pidx;
         pidx.reserve( npts_ );
 
-        for ( size_t ip = 0; ip < npts_; ++ip )
+        for ( unsigned int ip = 0; ip < npts_; ++ip )
             pidx.push_back(
                 PointIndex3::Value( PointIndex3::Point( coords( ip, 0 ), coords( ip, 1 ), coords( ip, 2 ) ), ip ) );
 
         tree_->build( pidx.begin(), pidx.end() );
     }
     else {
-        for ( size_t ip = 0; ip < npts_; ++ip )
+        for ( unsigned int ip = 0; ip < npts_; ++ip )
             tree_->insert(
                 PointIndex3::Value( PointIndex3::Point( coords( ip, 0 ), coords( ip, 1 ), coords( ip, 2 ) ), ip ) );
     }
